feat: add onboarding screen and empty out home screen
